requirements.txt
tensorflow=2.11.0
dask=2023.3.1
pandas=1.5.3
OpenEXR=1.3.9
waymo-open-dataset-tf-2-11-0==1.6.0
pillow=9.2.0
matplotlib= 3.6.1
opencv-python=4.8.0.74
open3d=0.16.0
easydict=1.11
ultralytics=8.0.145